\name{topGO-package}
\alias{topGO-package}
\alias{topGO}

\docType{package}

\title{Enrichment analysis for Gene Ontology}

\description{
  topGO package provides tools for testing GO terms while accounting
  for the topology of the GO graph. Different test statistics and
  different methods for eliminating local similarities and dependencies
  between GO terms can be implemented and applied.
}

\details{
  \tabular{ll}{
    Package: \tab topGO\cr
    Type: \tab Package\cr
    Version: \tab 1.0\cr
    Date: \tab 2006-10-02\cr
    License: \tab What license is it under?\cr
  }
  TODO: An overview of how to use the package, including the most important functions
}

\author{
  Adrian Alexa, J\"org Rahnenf\"uhrer
  
  Maintainer: Adrian Alexa
}

\references{
  Alexa A., Rahnenf\"uhrer J., Lengauer T., Improved scoring of
  functional groups from gene expression data by decorrelating GO
  graph structure, Bioinformatics 22(13): 1600-1607, 2006
}

\seealso{
  \code{\link{topGOdata-class}},  \code{\link{groupStats-class}},
  \code{\link{getSigGroups-methods}}
}


\keyword{package}
